feat(admin/backfill): extract job runner into dedicated component and preserve inputs on error
